backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Excel Processing
openpyxl==3.1.2

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
PyJWT==2.8.1

# AWS S3 (Optional)
boto3==1.34.20
botocore==1.34.20

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0
isort==5.13.2

# Monitoring & Logging
python-logging-loki==0.3.2
